+++ chunkflow/operators/save.py
@@ -1,3 +1,4 @@
+import time
 import os
 import json
 import numpy as np
@@ -51,16 +52,21 @@ class SaveOperator(OperatorBase):
  
  
     def __call__(self, chunk, log=None, output_bbox=None):
+        start = time.time()
+
         chunk_slices = chunk.slices 
         # transpose czyx to xyzc order
         arr = np.transpose(chunk)
         self.volume[chunk_slices[::-1]] = arr
 
-        if self.upload_log:
-            self._upload_log(log, output_bbox)
-
         if self.create_thumbnail:
             self._create_thumbnail(chunk)
+        
+        # add timer for save operation itself
+        elapsed = time.time() - start
+        log['timer'][self.name] = time.time() - start
+        if self.upload_log:
+            self._upload_log(log, output_bbox)
 
     def _create_thumbnail(self, chunk):
         if self.verbose:
@@ -88,6 +94,10 @@ class SaveOperator(OperatorBase):
     def _upload_log(self, log, output_bbox):
         assert log
         assert isinstance(output_bbox, Bbox)
+
+        if self.verbose:
+            print('uploaded log: ', log)
+
         # write to google cloud storage 
         self.log_storage.put_file(
             file_path=output_bbox.to_filename() + '.json',